Tests sin

+++ tinygrad/function.py
@@ -39,7 +39,7 @@ class Reciprocal(Function):
 class Sin(Function):
   def forward(self, x:LazyBuffer) -> LazyBuffer:
     # normalize x with analogue of math.fmod(x, 2*pi), the double cast is to get the floor
-    r = x.e(BinaryOps.DIV, x.const(math.pi*2))
+    r = x.cast(dtypes.float64).e(BinaryOps.DIV, x.const(math.pi*2))
     self.x = ret = acc = r.e(BinaryOps.SUB, r.cast(dtypes.int).cast(x.dtype)).e(BinaryOps.MUL, r.const(math.pi*2))
     self.precision = 14
     #if x.dtype == dtypes.float64:
